feat：“项目初始化，基础代码生成dto和vo”

+++ src/main/java/com/cool/pandora/model/dto/questionView/QuestionViewEditRequest.java
@@ -22,12 +22,12 @@ public class QuestionViewEditRequest implements Serializable {
     /**
      * 访问用户Id
      */
-    private Integer userId;
+    private Long userId;
 
     /**
      * 被浏览的面试题的唯一标识
      */
-    private Integer questionId;
+    private Long questionId;
 
     private static final long serialVersionUID = 1L;
 }
