+++ src/main/java/com/cool/pandora/model/entity/QuestionView.java
@@ -25,12 +25,12 @@ public class QuestionView implements Serializable {
     /**
      * 访问用户Id
      */
-    private Integer userId;
+    private Long userId;
 
     /**
      * 被浏览的面试题的唯一标识
      */
-    private Integer questionId;
+    private Long questionId;
 
     /**
      * 用户访问该题目的时间戳
